Bulk deletion looking good

+++ app/controllers/api/v1/plans_controller.rb
@@ -44,8 +44,12 @@ class Api::V1::PlansController < ApiController
 
   def destroy_items
     return permission_denied_error unless @plan && current_user.owns?(@plan)
-    marks = Mark.where( user_id: current_user.id, place_id: params[:place_ids] )
-    Item.where( plan_id: @plan.id, mark_id: marks.pluck(:id) ).destroy_all
+    if params[:item_ids].present?
+      @plan.items.where( id: params[:item_ids] ).destroy_all
+    elsif params[:place_ids].present?
+      marks = Mark.where( user_id: current_user.id, place_id: params[:place_ids] )
+      Item.where( plan_id: @plan.id, mark_id: marks.pluck(:id) ).destroy_all
+    end
     success
   end
